Improve on code duplication
