+++ src/components/Login/OryLoginForm.tsx
@@ -1,125 +1,19 @@
-import {
-    SelfServiceLoginFlow,
-    SubmitSelfServiceLoginFlowBody,
-    SubmitSelfServiceLoginFlowWithPasswordMethodBody as ValuesType,
-    SubmitSelfServiceSettingsFlowWithTotpMethodBody as TotpValuesType,
-} from "@ory/client";
-import { Alert, Form, message, Row } from "antd";
-import { AxiosError } from "axios";
+import { Alert, Form, Row } from "antd";
 import { useTranslation } from "next-i18next";
-import { useRouter } from "next/router";
-import React, { useEffect, useState } from "react";
+import React from "react";
 
-import { oryGetLoginFlow, orySubmitLogin } from "../../api/ory";
-import { ory } from "../../lib/orysdk";
-import { handleFlowError } from "../../lib/orysdk/errors";
-import { getUiNode } from "../../lib/orysdk/utils";
 import Input from "../AletheiaInput";
 import Button, { ButtonType } from "../Button";
 import InputPassword from "../InputPassword";
-import Loading from "../Loading";
 
-const OryLoginForm = () => {
-    const [flow, setFlow] = useState<SelfServiceLoginFlow>();
-    const [isLoading, setIsLoading] = useState(false);
+const OryLoginForm = ({
+    flow,
+    onFinish,
+    onFinishFailed,
+    isLoading,
+    onFinishTotp,
+}) => {
     const { t } = useTranslation();
-    const router = useRouter();
-
-    useEffect(() => {
-        oryGetLoginFlow({ router, setFlow, t });
-    }, []);
-
-    const onSubmit = (values: SubmitSelfServiceLoginFlowBody) => {
-        orySubmitLogin({ router, flow, setFlow, t, values }).then(() => {
-            setIsLoading(false);
-        });
-    };
-
-    const onSubmitTotp = (values: SubmitSelfServiceLoginFlowBody) =>
-        ory
-            .submitSelfServiceLoginFlow(String(flow?.id), undefined, values)
-            // We logged in successfully! Let's bring the user home.
-            .then(() => {
-                if (flow?.return_to) {
-                    window.location.href = flow?.return_to;
-                    return;
-                }
-                router.push("/");
-            })
-            .catch(handleFlowError(router, "login", setFlow, t))
-            .catch((err: AxiosError) => {
-                // If the previous handler did not catch the error it's most likely a form validation error
-                if (err.response?.status === 400) {
-                    // Yup, it is!
-                    setFlow(err.response?.data);
-                    return message.error(t("profile:totpIncorectCodeMessage"));
-                }
-
-                return Promise.reject(err);
-            });
-
-    if (!flow) {
-        return <Loading />;
-    }
-
-    let flowValues: ValuesType = {
-        csrf_token: "",
-        method: "password",
-        password: "",
-        password_identifier: "",
-        identifier: "",
-    };
-
-    let totpValues: TotpValuesType = {
-        csrf_token: "",
-        method: "totp",
-        totp_code: "",
-    };
-
-    const initializeCsrf = () => {
-        const csrfNode = getUiNode(flow, "name", "csrf_token");
-        if (csrfNode) {
-            flowValues.csrf_token = csrfNode.value;
-        }
-    };
-
-    const initializeCsrfTotp = () => {
-        const csrfNode = getUiNode(flow, "name", "csrf_token");
-        if (csrfNode) {
-            totpValues.csrf_token = csrfNode.value;
-        }
-    };
-
-    const onFinish = (values) => {
-        if (!isLoading) {
-            setIsLoading(true);
-            const { password, email } = values;
-            initializeCsrf();
-            flowValues = {
-                ...flowValues,
-                password,
-                password_identifier: email,
-            };
-            onSubmit(flowValues);
-        }
-    };
-
-    const onFinishTotp = (values) => {
-        initializeCsrfTotp();
-        totpValues = {
-            ...totpValues,
-            totp_code: values.totp,
-        };
-        onSubmitTotp(totpValues);
-    };
-
-    const onFinishFailed = (errorInfo) => {
-        if (typeof errorInfo === "string") {
-            message.error(errorInfo);
-        } else {
-            message.error(t("login:loginFailedMessage"));
-        }
-    };
 
     return (
         <>
